+++ src/Integration/ExceptionListenerIntegration.php
@@ -18,8 +18,9 @@ final class ExceptionListenerIntegration implements IntegrationInterface
      */
     public function setupOnce(): void
     {
-        ErrorHandler::addExceptionListener(static function (\Throwable $throwable): void {
-            Hub::getCurrent()->captureException($throwable);
+        $errorHandler = ErrorHandler::registerOnce(ErrorHandler::DEFAULT_RESERVED_MEMORY_SIZE, false);
+        $errorHandler->addExceptionHandlerListener(static function (\Throwable $exception): void {
+            Hub::getCurrent()->captureException($exception);
         });
     }
 }