+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 - Mark Sentry internal frames when using `attach_stacktrace` as `in_app` `false` (#786)
-- Increase default severity of `E_RECOVERABLE_ERROR` to `Severity::ERROR`, instead
-  of `Severity::WARNING` (#792)
+- Increase default severity of `E_RECOVERABLE_ERROR` to `Severity::ERROR`, instead of warning (#792)
+- Make it possible to register fatal error listeners separately from the error listeners
+  and change the type of the reported exception to `\Sentry\Exception\FatalErrorException` (#788)
 - Add a static factory method to create a breadcrumb from an array of data (#798)
 - Add support for `SENTRY_ENVRIONMENT` and `SENTRY_RELEASE` environment variables (#810)
 - Fix the default value of the `$exceptions` property of the Event class (#806)